was able to run my first test which failed but I got this going I guess
